refactor: Change plan-item directory
