+++ web/app/components/billing/pricing/plans/plan-item/index.tsx
@@ -2,11 +2,11 @@
 import type { FC } from 'react'
 import React, { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import type { BasicPlan } from '../../type'
-import { Plan } from '../../type'
-import { ALL_PLANS } from '../../config'
-import Toast from '../../../base/toast'
-import { PlanRange } from '../select-plan-range'
+import type { BasicPlan } from '../../../type'
+import { Plan } from '../../../type'
+import { ALL_PLANS } from '../../../config'
+import Toast from '../../../../base/toast'
+import { PlanRange } from '../../select-plan-range'
 import { useAppContext } from '@/context/app-context'
 import { fetchSubscriptionUrls } from '@/service/billing'
 import List from './list'
