+++ web/app/components/billing/pricing/plans/plan-item/list/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { type BasicPlan, Plan } from '../../../type'
+import { type BasicPlan, Plan } from '../../../../type'
 import Item from './item'
 import { useTranslation } from 'react-i18next'
-import { ALL_PLANS, NUM_INFINITE } from '../../../config'
+import { ALL_PLANS, NUM_INFINITE } from '../../../../config'
 import Divider from '@/app/components/base/divider'
 
 type ListProps = {
